Conflicts:
	gdal/ogr/ogrsf_frmts/gml/gmlhandler.cpp
	gdal/ogr/ogrsf_frmts/nas/nashandler.cpp

+++ gdal/ogr/ogrsf_frmts/nas/nashandler.cpp
@@ -605,7 +605,7 @@ void NASHandler::fatalError( const SAXParseException &exception)
     pszErrorMessage = tr_strdup( exception.getMessage() );
     CPLError( CE_Failure, CPLE_AppDefined,
               "XML Parsing Error: %s at line %d, column %d\n",
-              pszErrorMessage, exception.getLineNumber(), exception.getColumnNumber() );
+              pszErrorMessage, (int)exception.getLineNumber(), (int)exception.getColumnNumber() );
 
     CPLFree( pszErrorMessage );
 }
